export u64

+++ src/x64.rs
@@ -74,7 +74,7 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
     }
 
     /// Returns 'None' if should stop. Otherwise, returns delay as u32.
-    pub fn next_delay(&mut self, current_ms: Option<TimerInstantU64<TIMER_HZ_MILLIS>>) -> Option<u32> {
+    pub fn next_delay(&mut self, current_ms: Option<TimerInstantU64<TIMER_HZ_MILLIS>>) -> Option<u64> {
         if current_ms.is_none() && self.operating_mode == OperatingMode::Duration {
             return None;
         }
@@ -85,7 +85,7 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
     }
 
     /// Duration operating mode
-    fn next_delay_duration(&mut self, current_ms: TimerInstantU64<TIMER_HZ_MILLIS>) -> Option<u32> {
+    fn next_delay_duration(&mut self, current_ms: TimerInstantU64<TIMER_HZ_MILLIS>) -> Option<u64> {
         // If start time is None, we're at the start of the move. Set start time.
         if self.start_time_ms.is_none() {
             self.start_time_ms = Some(current_ms);
@@ -100,7 +100,7 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
             self.acceleration_steps += Fix::ONE;
             self.current_step += 1;
             self.current_delay = self.first_delay;
-            return Some(self.first_delay.to_num::<u32>());
+            return Some(self.first_delay.to_num::<u64>());
         }
 
         // If the time remaining is less than the time it took to accelerate, slow down.
@@ -108,7 +108,7 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
             let time_remaining = self.target_duration_ms - current_duration;
             if time_remaining <= *acceleration_duration_ms {
                 self.slow_down();
-                return Some(self.current_delay.to_num::<u32>());
+                return Some(self.current_delay.to_num::<u64>());
             }
         }
 
@@ -119,21 +119,21 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
                 self.acceleration_duration_ms = Some(current_duration);
             }
             self.current_step += 1;
-            Some(self.current_delay.to_num::<u32>())
+            Some(self.current_delay.to_num::<u64>())
         } else {
             self.speed_up();
-            Some(self.current_delay.to_num::<u32>())
+            Some(self.current_delay.to_num::<u64>())
         }
     }
 
     /// Step operating mode
-    fn next_delay_step(&mut self) -> Option<u32> {
+    fn next_delay_step(&mut self) -> Option<u64> {
         // If current step is 0, we're at the start of the move. Return the first delay and increase acceleration steps and current step.
         if self.current_step == Fix::ZERO {
             self.acceleration_steps += Fix::ONE;
             self.current_step += 1;
             self.current_delay = self.first_delay;
-            return Some(self.first_delay.to_num::<u32>());
+            return Some(self.first_delay.to_num::<u64>());
         }
 
         // If current step is bigger or equal to the target step, we're at the end of the move. Return None.
@@ -144,17 +144,17 @@ impl<const TIMER_HZ_MICROS: u32> Stepgen<TIMER_HZ_MICROS> {
         // If the current step is bigger are equal than the target step minus the acceleration steps, we need to slow down.
         if self.current_step >= self.target_step - self.acceleration_steps.to_num::<u64>() {
             self.slow_down();
-            return Some(self.current_delay.to_num::<u32>());
+            return Some(self.current_delay.to_num::<u64>());
         }
 
         // If the current delay is equal to the target delay, we're at the target speed. Return the current delay.
         // Else, we need to accelerate.
         if self.current_delay == self.target_delay {
             self.current_step += 1;
-            Some(self.current_delay.to_num::<u32>())
+            Some(self.current_delay.to_num::<u64>())
         } else {
             self.speed_up();
-            Some(self.current_delay.to_num::<u32>())
+            Some(self.current_delay.to_num::<u64>())
         }
     }
